+++ base/strings/substring.jl
@@ -37,20 +37,20 @@ struct SubString{T<:AbstractString} <: AbstractString
         return new(s, i-1, nextind(s,j)-i)
     end
 
-    global function unsafe_substring(s::T, first_index::Int, ncodeunits::Int) where {T <: AbstractString}
+    global function raw_substring(s::T, first_index::Int, ncodeunits::Int) where {T <: AbstractString}
         @boundscheck @inline checkbounds(codeunits(s), first_index:(first_index + n_codeunits - 1))
         new{T}(s, first_index - 1, ncodeunits)
     end
 
-    global function unsafe_substring(s::SubString{T}, first_index::Int, ncodeunits::Int) where {T <: AbstractString}
+    global function raw_substring(s::SubString{T}, first_index::Int, ncodeunits::Int) where {T <: AbstractString}
         @boundscheck @inline checkbounds(codeunits(s), first_index:(first_index + n_codeunits - 1))
         new{T}(s.string, first_index + s.offset - 1, n_codeunits)
     end
 end
 
 """
-    unsafe_substring(s::AbstractString, first_index::Int, n_codeunits::Int)::SubString{typeof(s)}
-    unsafe_substring(s::SubString{S}, first_index::Int, n_codeunits::Int)::SubString{S}
+    raw_substring(s::AbstractString, first_index::Int, n_codeunits::Int)::SubString{typeof(s)}
+    raw_substring(s::SubString{S}, first_index::Int, n_codeunits::Int)::SubString{S}
 
 Create a substring of `s` spanning the codeunits `first_index:(first_index + n_codeunits - 1)`.
 
@@ -69,25 +69,25 @@ substrings of these, but may not be permitted for custom subtypes of `AbstractSt
 ```jldoctest
 julia> s = "Hello, Bjørn!";
 
-julia> ss = unsafe_substring(s, 3, 10)
+julia> ss = raw_substring(s, 3, 10)
 "lo, Bjørn"
 
 julia> typeof(ss)
 SubString{String}
 
-julia> ss2 = unsafe_substring(ss, 2, 6)
+julia> ss2 = raw_substring(ss, 2, 6)
 "o, Bj\\xc3"
 
 julia> typeof(ss2)
 SubString{String}
 ```
 """
-function unsafe_substring end
+function raw_substring end
 
 
-function unsafe_substring(s::AbstractString, first_index::Int, n_codeunits::Int)
+function raw_substring(s::AbstractString, first_index::Int, n_codeunits::Int)
     @boundscheck @inline checkbounds(codeunits(s), first_index:(first_index + n_codeunits - 1))
-    return @inbounds unsafe_substring(s, first_index, n_codeunits)
+    return @inbounds raw_substring(s, first_index, n_codeunits)
 end
 
 @propagate_inbounds SubString(s::T, i::Int, j::Int) where {T<:AbstractString} = SubString{T}(s, i, j)
@@ -99,13 +99,13 @@ end
     SubString(s.string, s.offset+i, s.offset+j)
 end
 
-SubString(s::AbstractString) = @inbounds unsafe_substring(s, 1, Int(ncodeunits(s))::Int)
+SubString(s::AbstractString) = @inbounds raw_substring(s, 1, Int(ncodeunits(s))::Int)
 SubString(s::SubString) = s
 
 # Unlike the un-parameterized SubString constructor, this function must allow creating
 # e.g. a SubString{SubString{String}}, as this type is what the user may have explicitly
 # requested.
-SubString{T}(s::T) where {T<:AbstractString} = @inbounds unsafe_substring(s, 1, ncodeunits(s))
+SubString{T}(s::T) where {T<:AbstractString} = @inbounds raw_substring(s, 1, ncodeunits(s))
 
 @propagate_inbounds view(s::AbstractString, r::AbstractUnitRange{<:Integer}) = SubString(s, r)
 @propagate_inbounds maybeview(s::AbstractString, r::AbstractUnitRange{<:Integer}) = view(s, r)