+++ test/strings/basic.jl
@@ -233,17 +233,17 @@ end
         @test (@views (x[3], x[1:2], x[[1,4]])) == ('c', "ab", "ad")
     end
 
-    @testset "unsafe_substring" begin
+    @testset "raw_substring" begin
         s = "abcdefgøø"
-        @test unsafe_substring(s, 1, 11) == s
-        @test unsafe_substring(s, 1, 3) == "abc"
-        @test unsafe_substring(s, 3, 3) == "cde"
-        @test unsafe_substring(s, 5, 4) == String(codeunits(s)[5:8])
-        @test unsafe_substring(s, 1, 2) isa SubString{String}
-        @test unsafe_substring(unsafe_substring(s, 2, 8), 1, 3) isa SubString{String}
+        @test raw_substring(s, 1, 11) == s
+        @test raw_substring(s, 1, 3) == "abc"
+        @test raw_substring(s, 3, 3) == "cde"
+        @test raw_substring(s, 5, 4) == String(codeunits(s)[5:8])
+        @test raw_substring(s, 1, 2) isa SubString{String}
+        @test raw_substring(raw_substring(s, 2, 8), 1, 3) isa SubString{String}
 
-        @test_throws BoundsError unsafe_substring(s, 0, 2)
-        @test_throws BoundsError unsafe_substring(s, 2, 11)
+        @test_throws BoundsError raw_substring(s, 0, 2)
+        @test_throws BoundsError raw_substring(s, 2, 11)
     end
 end